# Memoise & cache the result of renderers, so we don't recalculate views multiple times.