--- a/render/benchmark_test.go
+++ b/render/benchmark_test.go
@@ -61,6 +61,7 @@ func benchmarkRender(b *testing.B, r render.Renderer) {
 	b.ResetTimer()
 
 	for i := 0; i < b.N; i++ {
+		render.ResetCache()
 		benchmarkRenderResult = r.Render(report)
 		if len(benchmarkRenderResult) == 0 {
 			b.Errorf("Rendered topology contained no nodes")
@@ -78,6 +79,7 @@ func benchmarkStats(b *testing.B, r render.Renderer) {
 
 	for i := 0; i < b.N; i++ {
 		// No way to tell if this was successful :(
+		render.ResetCache()
 		benchmarkStatsResult = r.Stats(report)
 	}
 }