--- a/test/reflect/deepequal.go
+++ b/test/reflect/deepequal.go
@@ -105,6 +105,9 @@ func deepValueEqual(v1, v2 reflect.Value, visited map[visit]bool, depth int) boo
 		return deepValueEqual(v1.Elem(), v2.Elem(), visited, depth+1)
 	case reflect.Struct:
 		for i, n := 0, v1.NumField(); i < n; i++ {
+			if v1.Type().Field(i).Tag.Get("deepequal") == "skip" {
+				continue
+			}
 			if !deepValueEqual(v1.Field(i), v2.Field(i), visited, depth+1) {
 				return false
 			}