Fixes #854

The idea is that to render hosts, you have to render processes, containers, pods etc.  All of these in turn render processes.  As the rendering pipeline is deterministic, ie rendering the same report will give you the same result, we can cache intermediate steps.   I introduce a random report ID and use reflection to get the address of the renderer, and use these two as a key for the cache.

Results in an 750% increase in performance of the host renderer, although this test might not be fair...

Before:
```
Toms-MacBook-Pro:render twilkie$ go test -bench . -- -bench-report-file report.json 
PASS
BenchmarkEndpointRender-8                	   10000	    133761 ns/op	   52781 B/op	     736 allocs/op
BenchmarkEndpointStats-8                 	100000000	        18.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessRender-8                 	    5000	    275635 ns/op	  114848 B/op	    1427 allocs/op
BenchmarkProcessStats-8                  	20000000	        65.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessWithContainerNameRender-8	    5000	    324229 ns/op	  131566 B/op	    1612 allocs/op
BenchmarkProcessWithContainerNameStats-8 	20000000	        75.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessNameRender-8             	    3000	    367554 ns/op	  146392 B/op	    1726 allocs/op
BenchmarkProcessNameStats-8              	100000000	        19.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerRender-8               	    2000	    521885 ns/op	  213154 B/op	    2540 allocs/op
BenchmarkContainerStats-8                	    3000	    433637 ns/op	  176031 B/op	    2255 allocs/op
BenchmarkContainerWithImageNameRender-8  	    2000	    554605 ns/op	  228969 B/op	    2707 allocs/op
BenchmarkContainerWithImageNameStats-8   	    3000	    432879 ns/op	  176003 B/op	    2255 allocs/op
BenchmarkContainerImageRender-8          	    2000	    686163 ns/op	  291427 B/op	    3150 allocs/op
BenchmarkContainerImageStats-8           	100000000	        18.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerHostnameRender-8       	    2000	    532886 ns/op	  220428 B/op	    2634 allocs/op
BenchmarkContainerHostnameStats-8        	100000000	        18.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkHostRender-8                    	    2000	    879954 ns/op	  373453 B/op	    4044 allocs/op
BenchmarkHostStats-8                     	10000000	       126 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodRender-8                     	    2000	    646618 ns/op	  268650 B/op	    3008 allocs/op
BenchmarkPodStats-8                      	100000000	        18.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodServiceRender-8              	    2000	    773566 ns/op	  326606 B/op	    3418 allocs/op
BenchmarkPodServiceStats-8               	100000000	        18.7 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/weaveworks/scope/render	34.022s
```

After:

```
Toms-MacBook-Pro:render twilkie$ go test -bench . -- -bench-report-file report.json 
PASS
BenchmarkEndpointRender-8                	   20000	     74163 ns/op	   29659 B/op	     377 allocs/op
BenchmarkEndpointStats-8                 	100000000	        10.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessRender-8                 	   30000	     44996 ns/op	   22172 B/op	     204 allocs/op
BenchmarkProcessStats-8                  	30000000	        39.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessWithContainerNameRender-8	   20000	     72938 ns/op	   31491 B/op	     319 allocs/op
BenchmarkProcessWithContainerNameStats-8 	20000000	        50.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkProcessNameRender-8             	  100000	     16413 ns/op	    7403 B/op	      76 allocs/op
BenchmarkProcessNameStats-8              	100000000	        10.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerRender-8               	   30000	     41609 ns/op	   21403 B/op	     135 allocs/op
BenchmarkContainerStats-8                	  100000	     10727 ns/op	    2969 B/op	      54 allocs/op
BenchmarkContainerWithImageNameRender-8  	   20000	     70158 ns/op	   32224 B/op	     236 allocs/op
BenchmarkContainerWithImageNameStats-8   	  100000	     10807 ns/op	    2969 B/op	      54 allocs/op
BenchmarkContainerImageRender-8          	  100000	     12298 ns/op	    5514 B/op	      62 allocs/op
BenchmarkContainerImageStats-8           	100000000	        10.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkContainerHostnameRender-8       	  200000	      8771 ns/op	    3657 B/op	      50 allocs/op
BenchmarkContainerHostnameStats-8        	100000000	        10.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkHostRender-8                    	   10000	    116569 ns/op	   59922 B/op	     327 allocs/op
BenchmarkHostStats-8                     	20000000	        80.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodRender-8                     	  100000	     12679 ns/op	    5514 B/op	      62 allocs/op
BenchmarkPodStats-8                      	100000000	        10.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkPodServiceRender-8              	  100000	     11310 ns/op	    4634 B/op	      59 allocs/op
BenchmarkPodServiceStats-8               	100000000	        10.3 ns/op	       0 B/op	       0 allocs/op
ok  	github.com/weaveworks/scope/render	31.782s
```