@tomwilkie 2016-01-23
This is going to conflict with #838 as I copied the DeepEqual implementation from there.  Put #838 in first, I don't mind rebasing this.