@paulbellamy 2016-01-25
Would passing around a `*report.Report` allow us to use the pointer as the cache key and skip the `report.Report.ID` thing?