@paulbellamy 2016-01-25
Overall, a lot better (cleaner and less complicated), than I was expecting. I think this could work.

Would probably be good to add a `render.PurgeCache()` func, for the tests and benchmarks.