@tomwilkie 2016-01-25
Roger. Could do report pointers, I wasn't keen on the id stuff myself. Will
try it out.

On Monday, 25 January 2016, Paul Bellamy <notifications@github.com> wrote:

> Overall, a lot better (cleaner and less complicated), than I was
> expecting. I think this could work.
>
> Would probably be good to add a render.PurgeCache() func, for the tests
> and benchmarks.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/weaveworks/scope/pull/851#issuecomment-174492641>.
>