@paulbellamy 2016-01-25
A couple potential suggestions, but with a `render.PurgeCache()` called in the benchmarks/tests, it LGTM.