@tomwilkie 2016-01-25
After adding the PurgeCache method between benchmark runs, it becomes obvious this isn't much of a performance increase (613703 -> 582766 ns/op, so within the margin).  On inspection, this is because there is actually very little work done twice by the host renderer, or any renderer.

However this does make a huge difference to the calls to /api/topology (from 7197010 -> 1447692 ns/op, or 5x faster).  As we hit this every second, this results in a reduction in CPU usage for the scope container from about ~25% to ~20% on my VM running the example app.

Still too high.