@tomwilkie 2016-01-26
>  If you're saying that the CPU usage is too high, then I agree

Yes thats what I meant.  I see the app using mostly 3% CPU, occasionally spiking to 20%.

> Actually, this should mean that we only have to run each renderer each time a probe report comes in, so we're at most running them 8ms/15seconds or 0.05% of the time. 

This won't happen as every http request creates a new report (with a new ID) so won't hit the cache - ie the cache is only used for things renderer multiple times in the same request.  We could improve the collector such that the same report is used across requests?