@paulbellamy 2016-01-26
> We could [have the collector cache the last report until a new one comes in]

:+1: