@tomwilkie 2016-01-27
Caching them also didn't make a huge different, but the app uses 0% CPU for a bunch of time, and occasionally spikes to 10ish.  I think this is fine for the app now.