@paulbellamy 2016-01-27
Why not `murmur3.New128`?